requirements.txt
# Runtime dependencies for OGM_slave_pi tools and daemon.
PyYAML>=6.0
pylibmodbus>=0.5.0  # Optional: libmodbus-backed RTU server binding.
